+++ CHANGELOG.md
@@ -80,12 +80,17 @@ flowchart LR
 - BREAKING: change `eval_frequency` default value to None to avoid confusion with hidden default value (#91)
 - BREAKING: rename Algo to Function ([#82](https://github.com/Substra/substrafl/pull/82))
 - BREAKING: clarify `EvaluationStrategy` arguments: change `rounds` to `eval_frequency` and `eval_rounds` (#85)
+- replace `schemas.xxx` by `substra.schemas.xxx` ([#105](https://github.com/Substra/substrafl/pull/105))
 
 ### Fixed
 
 - BREAKING: Given local code dependencies are now copied to the level of the running script systematically ([#99](https://github.com/Substra/substrafl/pull/99/files))
 - Docker images are pruned in main check of Github Action to free disk space while test run ([#102](https://github.com/Substra/substrafl/pull/102))
 - Pass `aggregation_lr` to the parent class for Scaffold. Fix [issue 103](https://github.com/Substra/substrafl/issues/103) ([#104](https://github.com/Substra/substrafl/pull/104))
+
+### Removed
+
+- `from substra import schemas` in `aggregation_node.py`, `test_data_node.py` and `train_data_node.py` ([#105](https://github.com/Substra/substrafl/pull/105))
 
 ## [0.34.0](https://github.com/Substra/substrafl/releases/tag/0.34.0) - 2023-02-20